+++ src/__tests__/rendererIntegrationTests/renderer-integrationtest.js
@@ -70,15 +70,16 @@ describe('full application integration tests', () => {
     await expect(app.getClientFlags(0)).resolves.toEqual(flags);
   });
 
-  doTest("loads flags through Electron's net module", async (fakeLD, app) => {
+  doTest('loads flags through a custom Electron net request', async (fakeLD, app) => {
     const env = fakeLD.addEnvironment(defaultEnvId);
     const flags = { flag1: 'value1', flag2: 'value2' };
     env.addUser(defaultUserKey, flags);
 
-    await app.start({ useNetModule: true });
+    await app.start({ httpRequest: true });
 
     await expect(app.getClientUserKey(0)).resolves.toEqual(defaultUserKey);
     await expect(app.getClientFlags(0)).resolves.toEqual(flags);
+    expect(env.getLastPollRequest().headers['x-test-http-request']).toEqual('electron.net');
   });
 
   doTest('sends events from renderer client', async (fakeLD, app) => {